# ms-moonlight-main/plugins/picpay-argocd-backend/docs/swagger.yaml
basePath: /api/argocd
host: moonlight.qa.ppay.me
info:
  contact:
    name: 'squad-atlantis'
    url: http://moonlight.limbo.work
  description: ArgoCD Endpoint BFF
  termsOfService: http://swagger.io/terms/
  title: moonlight-plugin-argocd
  version: '1.0'
paths:
  /clusters:
    get:
      consumes:
        - '*/*'
      description: Returns Clusters.
      produces:
        - application/json
      responses:
        '500':
          description: Internal Server Error
          schema:
            type: object
            properties:
              message:
                type: string
        '200':
          description: OK
          schema:
            properties:
              name:
                type: string
              bu:
                type: string
            description: List of Clusters available on ArgoCD
            type: object
      summary: ArgoCD Listed Clusters
      parameters:
        - in: query
          name: bu
          required: false
          type: string
          description: Returns all clusters from specific Business Unit
      tags:
        - clusters
  /clusters/{applicationName}:
    get:
      consumes:
        - '*/*'
      description: Returns Clusters for an Application.
      produces:
        - application/json
      responses:
        '500':
          description: Internal Server Error
          schema:
            type: object
            properties:
              message:
                type: string
        '200':
          description: OK
          schema:
            properties:
              name:
                type: string
              bu:
                type: string
            description: List of Clusters available on ArgoCD
            type: object
      summary: ArgoCD Listed Clusters
      parameters:
        - in: path
          name: applicationName
          required: true
          type: string
          description: The name or repository of the application
      tags:
        - clusters
swagger: '2.0'